+++ docs/changelog.md
@@ -6,6 +6,7 @@ Change Log
 * Fix: Enum classes that only have an init block now also generate the required semicolon (#1952)
 * Fix: Preserve typealiases in `KSAnnotation.toAnnotationSpec()`. (#1945)
 * Fix: Preserve nullability in `KSType.toClassName()`. (#1956)
+* Fix: Return `FunSpec.Builder` from `FunSpec.Builder.jvmModifiers()` to not break chaining. (#2004)
 * New: Add `KSTypeAlias.toClassName()`. (#1956)
 * New: Add `KSType.toClassNameOrNull()`. (#1956)